extract LineVariant and ManuscriptLine into separate files
