+++ ebl/corpus/application/alignment_updater.py
@@ -6,7 +6,9 @@ import attr
 from ebl.corpus.application.chapter_updater import ChapterUpdater
 from ebl.corpus.domain.alignment import Alignment, ManuscriptLineAlignment
 from ebl.corpus.domain.chapter import Chapter, ChapterItem
-from ebl.corpus.domain.line import Line, LineVariant, ManuscriptLine
+from ebl.corpus.domain.line import Line
+from ebl.corpus.domain.manuscript_line import ManuscriptLine
+from ebl.corpus.domain.line_variant import LineVariant
 from ebl.transliteration.domain.alignment import AlignmentError
