+++ ebl/corpus/application/schemas.py
@@ -14,7 +14,9 @@ from ebl.corpus.domain.chapter import (
     Chapter,
     Classification,
 )
-from ebl.corpus.domain.line import Line, LineVariant, ManuscriptLine
+from ebl.corpus.domain.line import Line
+from ebl.corpus.domain.manuscript_line import ManuscriptLine
+from ebl.corpus.domain.line_variant import LineVariant
 from ebl.corpus.domain.manuscript import (
     Manuscript,
     ManuscriptType,
