+++ ebl/corpus/domain/chapter_transformer.py
@@ -2,7 +2,9 @@ from typing import Iterable
 
 from lark.visitors import v_args
 
-from ebl.corpus.domain.line import Line, LineVariant, ManuscriptLine
+from ebl.corpus.domain.line import Line
+from ebl.corpus.domain.manuscript_line import ManuscriptLine
+from ebl.corpus.domain.line_variant import LineVariant
 from ebl.corpus.domain.manuscript import (
     Manuscript,
     Period,
