+++ ebl/corpus/domain/line.py
@@ -1,251 +1,14 @@
-from functools import singledispatch
-from typing import Optional, Sequence, Set, Tuple, Union, cast, Callable
+from typing import Optional, Sequence
 
 import attr
 import pydash
+from ebl.corpus.domain.line_variant import LineVariant
+from ebl.corpus.domain.manuscript_line import ManuscriptLine, ManuscriptLineLabel
 
-from ebl.corpus.domain.create_alignment_map import create_alignment_map
-from ebl.corpus.domain.enclosure_validator import validate
 from ebl.merger import Merger
-from ebl.transliteration.domain.dollar_line import DollarLine
-from ebl.transliteration.domain.enclosure_visitor import set_enclosure_type
-from ebl.transliteration.domain.label_validator import validate_labels
-from ebl.transliteration.domain.labels import Label
-from ebl.transliteration.domain.language_visitor import set_language
-from ebl.transliteration.domain.line import EmptyLine
 from ebl.transliteration.domain.line_number import AbstractLineNumber, OldLineNumber
-from ebl.transliteration.domain.markup import MarkupPart
-from ebl.transliteration.domain.note_line import NoteLine
-from ebl.transliteration.domain.parallel_line import ParallelLine
-from ebl.transliteration.domain.text_line import AlignmentMap, TextLine, merge_tokens
-from ebl.transliteration.domain.tokens import Token
+from ebl.transliteration.domain.text_line import TextLine
 from ebl.transliteration.domain.translation_line import TranslationLine
-from ebl.transliteration.domain.word_tokens import AbstractWord
-from ebl.corpus.domain.manuscript import Manuscript
-from ebl.transliteration.domain.atf import Atf
-
-
-ManuscriptLineLabel = Tuple[int, Sequence[Label], AbstractLineNumber]
-
-
-@attr.s(auto_attribs=True, frozen=True)
-class ManuscriptLine:
-    manuscript_id: int
-    labels: Sequence[Label] = attr.ib(validator=validate_labels)
-    line: Union[TextLine, EmptyLine]
-    paratext: Sequence[Union[DollarLine, NoteLine]] = tuple()
-    omitted_words: Sequence[int] = tuple()
-
-    @property
-    def label(self) -> Optional[ManuscriptLineLabel]:
-        return (
-            (self.manuscript_id, self.labels, self.line.line_number)
-            if isinstance(self.line, TextLine)
-            else None
-        )
-
-    @property
-    def line_prefix_atf(self) -> Atf:
-        return (
-            Atf(
-                " ".join(
-                    label_element.to_value()
-                    for label_element in self.labels
-                    if hasattr(label_element, "to_value")
-                )
-            )
-            if self.labels
-            else Atf("")
-        )
-
-    @property
-    def is_beginning_of_side(self) -> bool:
-        return (
-            self.line.line_number.is_beginning_of_side
-            if isinstance(self.line, TextLine)
-            else False
-        )
-
-    @property
-    def is_end_of_side(self) -> bool:
-        return any(
-            line.is_end_of for line in self.paratext if isinstance(line, DollarLine)
-        )
-
-    @property
-    def is_empty(self) -> bool:
-        return isinstance(self.line, EmptyLine)
-
-    def get_atf(self, get_manuscript: Callable[[int], Manuscript]) -> Atf:
-        if self.is_empty:
-            return Atf("")
-        siglum = get_manuscript(self.manuscript_id).siglum
-        line_prefix = f"{self.line_prefix_atf} " if self.line_prefix_atf else ""
-        paratext = "\n".join([paratext.atf for paratext in self.paratext])
-        paratext = f"\n{paratext}" if self.paratext else ""
-        return Atf(f"{siglum} {line_prefix}{self.line.atf}{paratext}")
-
-    def merge(self, other: "ManuscriptLine") -> "ManuscriptLine":
-        merged_line = self.line.merge(other.line)
-        return attr.evolve(other, line=merged_line)
-
-    def update_alignments(self, aligment_map: AlignmentMap) -> "ManuscriptLine":
-        return attr.evolve(
-            self,
-            line=self.line.update_alignments(aligment_map),
-            omitted_words=tuple(
-                aligment_map[index]
-                for index in self.omitted_words
-                if index < len(aligment_map) and aligment_map[index] is not None
-            ),
-        )
-
-    def get_line_content(self) -> Sequence[Token]:
-        return tuple() if self.is_empty else cast(TextLine, self.line).content
-
-
-@attr.s(auto_attribs=True, frozen=True)
-class LineVariant:
-    reconstruction: Sequence[Token] = attr.ib(
-        converter=pydash.flow(set_enclosure_type, set_language)
-    )
-    note: Optional[NoteLine] = None
-    manuscripts: Sequence[ManuscriptLine] = tuple()
-    parallel_lines: Sequence[ParallelLine] = tuple()
-    intertext: Sequence[MarkupPart] = tuple()
-
-    @reconstruction.validator
-    def validate_reconstruction(self, _, value):
-        validate(value)
-
-    @property
-    def reconstruction_atf(self) -> Atf:
-        return Atf(" ".join([token.value for token in self.reconstruction]))
-
-    @property
-    def parallels_atf(self) -> Atf:
-        return (
-            Atf("\n".join(parallel.atf for parallel in self.parallel_lines))
-            if self.parallel_lines
-            else Atf("")
-        )
-
-    def get_manuscript_lines_atf(
-        self, get_manuscript: Callable[[int], Manuscript]
-    ) -> Atf:
-        atf_lines = [
-            manuscript_line.get_atf(get_manuscript)
-            for manuscript_line in self.manuscripts
-        ]
-        return Atf("\n".join(atf_line for atf_line in atf_lines if atf_line))
-
-    @property
-    def manuscript_ids(self) -> Sequence[int]:
-        return [manuscript.manuscript_id for manuscript in self.manuscripts]
-
-    @property
-    def manuscript_line_labels(self) -> Sequence[ManuscriptLineLabel]:
-        return [
-            manuscript_line.label
-            for manuscript_line in self.manuscripts
-            if manuscript_line.label
-        ]
-
-    @property
-    def _variant_alignments(self) -> Set[Optional[int]]:
-        return {
-            manuscript_token.alignment
-            for manuscript in self.manuscripts
-            for manuscript_token in manuscript.get_line_content()
-            if isinstance(manuscript_token, AbstractWord)
-            if manuscript_token.has_variant
-        }
-
-    @property
-    def _omitted_words(self) -> Set[Optional[int]]:
-        return {
-            index
-            for manuscript in self.manuscripts
-            for index in manuscript.omitted_words
-        }
-
-    def get_manuscript_line(self, manuscript_id: int) -> Optional[ManuscriptLine]:
-        return (
-            pydash.chain(self.manuscripts)
-            .filter(lambda manuscript: manuscript.manuscript_id == manuscript_id)
-            .head()
-            .value()
-        )
-
-    def get_manuscript_text_line(self, manuscript_id: int) -> Optional[TextLine]:
-        return (
-            pydash.chain(self.manuscripts)
-            .filter(lambda manuscript: manuscript.manuscript_id == manuscript_id)
-            .map_(lambda manuscript: manuscript.line)
-            .filter(lambda line: isinstance(line, TextLine))
-            .head()
-            .value()
-        )
-
-    def merge(self, other: "LineVariant") -> "LineVariant":
-        merged_reconstruction = merge_tokens(self.reconstruction, other.reconstruction)
-        alignment_map = create_alignment_map(self.reconstruction, merged_reconstruction)
-
-        def merge_manuscript(
-            old: ManuscriptLine, new: ManuscriptLine
-        ) -> ManuscriptLine:
-            return old.merge(new).update_alignments(alignment_map)
-
-        merged_manuscripts = Merger(repr, merge_manuscript).merge(
-            self.manuscripts, other.manuscripts
-        )
-        return (
-            attr.evolve(
-                other,
-                reconstruction=merged_reconstruction,
-                manuscripts=tuple(merged_manuscripts),
-            )
-            .set_has_variant_aligment()
-            .set_has_omitted_aligment()
-        )
-
-    def set_has_variant_aligment(self) -> "LineVariant":
-        variant_alignments = self._variant_alignments
-
-        @singledispatch
-        def set_flag(token: Token, index: int) -> Token:
-            return token
-
-        @set_flag.register(AbstractWord)
-        def _(token: AbstractWord, index: int) -> AbstractWord:
-            return token.set_has_variant_alignment(index in variant_alignments)
-
-        return attr.evolve(
-            self,
-            reconstruction=tuple(
-                set_flag(token, index)
-                for index, token in enumerate(self.reconstruction)
-            ),
-        )
-
-    def set_has_omitted_aligment(self) -> "LineVariant":
-        omitted_alignments = self._omitted_words
-
-        @singledispatch
-        def set_flag(token: Token, index: int) -> Token:
-            return token
-
-        @set_flag.register(AbstractWord)
-        def _(token: AbstractWord, index: int) -> AbstractWord:
-            return token.set_has_omitted_alignment(index in omitted_alignments)
-
-        return attr.evolve(
-            self,
-            reconstruction=tuple(
-                set_flag(token, index)
-                for index, token in enumerate(self.reconstruction)
-            ),
-        )
 
 
 @attr.s(auto_attribs=True, frozen=True)
