+++ ebl/corpus/web/chapter_schemas.py
@@ -12,7 +12,9 @@ from ebl.corpus.application.schemas import (
     labels,
     manuscript_id,
 )
-from ebl.corpus.domain.line import Line, LineVariant, ManuscriptLine
+from ebl.corpus.domain.line import Line
+from ebl.corpus.domain.manuscript_line import ManuscriptLine
+from ebl.corpus.domain.line_variant import LineVariant
 from ebl.corpus.domain.parser import parse_paratext
 from ebl.fragmentarium.application.joins_schema import JoinsSchema
 from ebl.fragmentarium.domain.joins import Joins
