+++ ebl/corpus/web/display_schemas.py
@@ -2,7 +2,9 @@ from typing import Union, Sequence
 import attr
 from marshmallow import Schema, fields
 from ebl.bibliography.domain.reference import Reference
-from ebl.corpus.domain.line import Line, ManuscriptLine, LineVariant
+from ebl.corpus.domain.line import Line
+from ebl.corpus.domain.manuscript_line import ManuscriptLine
+from ebl.corpus.domain.line_variant import LineVariant
 from ebl.corpus.domain.manuscript import Manuscript, OldSiglum
 from ebl.fragmentarium.domain.joins import Join
 from ebl.transliteration.application.museum_number_schema import MuseumNumberSchema
