+++ ebl/tests/corpus/test_extant_lines_schema.py
@@ -1,5 +1,7 @@
 from ebl.corpus.domain.chapter import Chapter
-from ebl.corpus.domain.line import Line, LineVariant, ManuscriptLine
+from ebl.corpus.domain.line import Line
+from ebl.corpus.domain.manuscript_line import ManuscriptLine
+from ebl.corpus.domain.line_variant import LineVariant
 from ebl.corpus.domain.manuscript import Manuscript
 from ebl.transliteration.domain.text_id import TextId
 from ebl.transliteration.domain.line_number import LineNumber
