+++ ebl/tests/factories/corpus.py
@@ -2,7 +2,9 @@ import factory.fuzzy
 import pydash
 
 from ebl.corpus.domain.chapter import Chapter, Classification
-from ebl.corpus.domain.line import Line, LineVariant, ManuscriptLine
+from ebl.corpus.domain.line import Line
+from ebl.corpus.domain.manuscript_line import ManuscriptLine
+from ebl.corpus.domain.line_variant import LineVariant
 from ebl.corpus.domain.manuscript import (
     Manuscript,
     ManuscriptType,
